% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseUtils.R
\name{getModelFitStat}
\alias{getModelFitStat}
\title{Fit Model Statistics}
\usage{
getModelFitStat(
  metric = c("AIC", "AICc", "BIC", "CP", "HQ", "HQc", "Rsq", "adjRsq", "SBC", "IC(3/2)",
    "IC(1)"),
  fit,
  type = c("linear", "logit", "poisson", "cox"),
  sigma_value
)
}
\arguments{
\item{metric}{Information criteria, including AIC, AICc, BIC, CP, HQ, HQc, Rsq, adjRsq and SBC}

\item{fit}{Object of linear model or general linear model}

\item{type}{"linear", "cox", "logit",or "poisson": to calculate information criteria value; for "linear", the "Least Square" method will be used; for others, "Maximum Likelyhood" method will be used.}
}
\description{
Fit Model Statistics with least square or likelihood method to return an information criteria value
}
