% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseUtils.R
\name{getXname}
\alias{getXname}
\title{Stepwise helper functions}
\usage{
getXname(formula, data)
}
\value{
data.frame
}
\description{
Extract information of parameters from stepwise regression.
}
\author{
Junhui Li, Kai Hu
}
