% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepRegShinyApp.R
\name{StepRegShinyApp}
\alias{StepRegShinyApp}
\title{StepReg Shiny App}
\usage{
StepRegShinyApp()
}
\description{
Launch the StepReg Shiny App for stepwise regression. In Step 1, upload your dataset, and in Step 2, choose the desired arguments to perform stepwise regression.
}
