% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseCox.R
\name{stepwiseCox}
\alias{stepwiseCox}
\title{Stepwise Cox Proportional Hazards Regression}
\usage{
stepwiseCox(
  formula,
  data,
  include = NULL,
  strategy = c("forward", "backward", "bidirection", "subset"),
  metric = c("SL", "AIC", "AICc", "SBC", "HQ", "HQc", "IC(3/2)", "IC(1)"),
  sle = 0.15,
  sls = 0.15,
  test_method_cox = c("efron", "breslow", "exact"),
  weights = NULL,
  best_n = NULL
)
}
\description{
Stepwise Cox regression analysis selects model based on information criteria and significant test with 'forward', 'backward', 'bidirection' and 'subset' variable selection method.
}
