% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseLinear.R
\name{stepwiseLinear}
\alias{stepwiseLinear}
\title{Stepwise Linear Model Regression}
\usage{
stepwiseLinear(
  formula,
  data,
  include = NULL,
  strategy = c("forward", "backward", "bidirection", "subset"),
  metric = c("AIC", "AICc", "BIC", "CP", "HQ", "HQc", "Rsq", "adjRsq", "SL", "SBC"),
  sle = 0.15,
  sls = 0.15,
  test_method_linear = c("Pillai", "Wilks", "Hotelling-Lawley", "Roy"),
  weights = NULL,
  best_n = Inf
)
}
\description{
Stepwise linear regression analysis selects model based on information criteria and F or approximate F test with 'forward', 'backward', 'bidirection' and 'subset' model strategy method.
}
