% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseLogit.R
\name{stepwiseLogit}
\alias{stepwiseLogit}
\title{Stepwise Logistic Regression}
\usage{
stepwiseLogit(
  formula,
  data,
  include = NULL,
  strategy = c("forward", "backward", "bidirection", "subset"),
  metric = c("SL", "AIC", "AICc", "SBC", "HQ", "HQc", "IC(3/2)", "IC(1)"),
  sle = 0.15,
  sls = 0.15,
  test_method_logit = c("Rao", "LRT"),
  weights = NULL,
  best_n = NULL
)
}
\description{
Stepwise logistic regression analysis selects model based on information criteria and Wald or Score test with 'forward', 'backward', 'bidirection' and 'subset' model strategy method.
}
