% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobacco.R
\docType{data}
\name{tobacco}
\alias{tobacco}
\title{tobacco}
\format{
A data frame containing 25 observations on 9 variables.
}
\usage{
data(tobacco)
}
\description{
data on chemical components of 25 tobacco leaf
}
\details{
\itemize{
  \item cigarette Rate of cigarette burn in inches per 1000 seconds.
  \item sugar Percent sugar in the leaf.
  \item nicotine Percent nicotine.
  \item nitrogen Percentage of nitrogen.
  \item chlorine Percentage of chlorine.
  \item potassium Percentage of potassium.
  \item phosphorus Percentage of phosphorus
  \item calcium Factor. Percentage of calcium.
  \item magnesium Percentage of magnesium.
}
}
\references{
Anderson, R. L. and Bancroft, T. A. (1952), Statistical Theory in Research, McGraw-Hill Book Company, Inc., New York, NY.
}
\keyword{datasets}
