% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseUtils.R
\name{getXname}
\alias{getXname}
\title{Stepwise helper functions}
\usage{
getXname(formula, data)
}
\arguments{
\item{formula}{(formula) The formula used for model fitting. The formula takes the form of a '~' (tilde) symbol, with the response variable(s) on the left-hand side, and the predictor variable(s) on the right-hand side. The 'lm()' function uses this formula to fit a regression model. A formula can be as simple as 'y ~ x'. For multiple predictors, they must be separated by the '+' (plus) symbol, e.g. 'y ~ x1 + x2'. To include an interaction term between variables, use the ':' (colon) symbol: 'y ~ x1 + x1:x2'. Use the '.' (dot) symbol to indicate that all other variables in the dataset should be included as predictors, e.g. 'y ~ .'. In the case of multiple response variables (multivariate), the formula can be specified as 'cbind(y1, y2) ~ x1 + x2'. By default, an intercept term is always included in the models, to exclude it, include '0' or '- 1' in your formula: 'y ~ 0 + x1', 'y ~ x1 + 0', and 'y ~ x1 - 1'.}

\item{data}{(data.frame) A dataset consisting of predictor variable(s) and response variable(s).}
}
\description{
Extract information of parameters from stepwise regression.
}
\author{
Junhui Li, Kai Hu
}
