% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.StepReg.R
\name{plot.StepReg}
\alias{plot.StepReg}
\title{Plots from a StepReg object}
\usage{
\method{plot}{StepReg}(x, ...)
}
\arguments{
\item{x}{StepReg object}

\item{...}{Not used}
}
\value{
line and point plot for summary of selection process.
}
\description{
plot.StepReg visualizes the variable selection procedure using a StepReg object
}
\examples{
\dontrun{
data(mtcars)
formula <- mpg ~ . + 0
p <- stepwise(formula = formula,
              data = mtcars,
              type = "linear",
              strategy = c("forward","bidirection","subset"),
              metric = c("AIC", "BIC"),
              best_n = 3)
plot(p)
}
}
