% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.StepReg.R
\name{plot.StepReg}
\alias{plot.StepReg}
\title{Plots from a StepReg object}
\usage{
\method{plot}{StepReg}(x, ...)
}
\arguments{
\item{x}{StepReg object}

\item{...}{further paramters}
}
\value{
formatted dataframe
}
\description{
print.StepReg prints to console the from an object of class StepReg
}
