% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.StepReg.R
\name{print.StepReg}
\alias{print.StepReg}
\title{Prints from a StepReg object}
\usage{
\method{print}{StepReg}(x, ...)
}
\arguments{
\item{x}{StepReg object}

\item{...}{further parameters}
}
\value{
only print some dataframe
}
\description{
print.StepReg prints to console the from an object of class StepReg
}
