% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remission.R
\docType{data}
\name{remission}
\alias{remission}
\title{remission}
\format{
A data frame with 27 rows and 7 columns.
}
\usage{
data(remission)
}
\description{
A dataset containing the remiss and 6 risk factors thought to be related to cancer remission.
}
\details{
\itemize{
  \item Remiss Indicates whether cancer remission occurred. A value of 1 indicates occurrence, while 0 indicates non-occurrence.
  \item cell
  \item smear
  \item infil
  \item li A measure of cell activity.
  \item blast
  \item temp
}
}
\references{
Lee, E. T. (1974). “A Computer Program for Linear Logistic Regression Analysis.” Computer Programs in Biomedicine 4:80–92.
}
\keyword{datasets}
