% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Plots from a StepReg object}
\usage{
report(x, report_name, format = c("html", "docx", "rtf", "pptx", "xlsx"))
}
\arguments{
\item{x}{StepReg object}

\item{report_name}{report name}

\item{format}{the format of report, choose one or more from 'html', 'docx', 'rtf', 'pptx', 'xlsx'. default is 'html'}
}
\description{
plot.StepReg visualizes the variable selection procedure using a StepReg object
}
\examples{
data(mtcars)
mtcars$yes <- mtcars$wt
formula <- mpg ~ . + 0
x <- stepwise(formula = formula,
              data = mtcars,
              type = "linear",
              strategy = "bidirection",
              metric = c("AIC", "BIC"))
report(x,report_name = "report", format = c("html","docx"))
}
