\name{StepReg-package}
\alias{StepReg-package}
\alias{StepReg}
\docType{package}
\title{
  Stepwise Regression Analysis
}
\description{
  	Stepwise regression analysis for variable selection can be used to get the best candidate final regression model in univariate or multivariate regression analysis with the 'forward', 'backward' and 'bidirection' steps. And best subset selection is also inclueded in this package. Procedure uses Akaike information criterion, corrected Akaike information criterion, Bayesian information criterion, Hannan and Quinn information criterion, corrected Hannan and Quinn information criterion, Schwarz criterion and significance levels as selection criteria. Multicollinearity detection in regression model are performed by checking tolerance value. Continuous variables nested within class effect and weighted stepwise regression are also considered.
}
\details{
\tabular{ll}{
Package: \tab StepReg\cr
Type: \tab Package\cr
Version: \tab 1.3.3\cr
Date: \tab 2019-11-25\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Junhui Li, Xiaohuan Lu, Kun Cheng,Wenxin Liu

Maintainer: Junhui Li <junhuili@cau.edu.cn>
}
\references{
Alsubaihi, A. A., Leeuw, J. D., and Zeileis, A. (2002). Variable selection in multivariable regression using sas/iml. , 07(i12).

Darlington, R. B. (1968). Multiple regression in psychological research and practice. Psychological Bulletin, 69(3), 161.

Hannan, E. J., & Quinn, B. G. (1979). The determination of the order of an autoregression. Journal of the Royal Statistical Society, 41(2), 190-195.

Harold Hotelling. (1992). The Generalization of Student's Ratio. Breakthroughs in Statistics. Springer New York.

Hocking, R. R. (1976). A biometrics invited paper. the analysis and selection of variables in linear regression. Biometrics, 32(1), 1-49.

Hurvich, C. M., & Tsai, C. (1989). Regression and time series model selection in small samples. Biometrika, 76(2), 297-307.

Judge, & GeorgeG. (1985). The Theory and practice of econometrics /-2nd ed. The Theory and practice of econometrics /. Wiley.

Mallows, C. L. (1973). Some comments on cp. Technometrics, 15(4), 661-676.

Mardia, K. V., Kent, J. T., & Bibby, J. M. (1979). Multivariate analysis. Mathematical Gazette, 37(1), 123-131.

Mckeon, J. J. (1974). F approximations to the distribution of hotelling's t20. Biometrika, 61(2), 381-383.

Mcquarrie, A. D. R., & Tsai, C. L. (1998). Regression and Time Series Model Selection. Regression and time series model selection /. World Scientific.

Pillai, K. C. S. (2006). Pillai's Trace. Encyclopedia of Statistical Sciences. John Wiley & Sons, Inc.

R.S. Sparks, W. Zucchini, & D. Coutsourides. (1985). On variable selection in multivariate regression. Communication in Statistics- Theory and Methods, 14(7), 1569-1587.

Sawa, T. (1978). Information criteria for discriminating among alternative regression models. Econometrica, 46(6), 1273-1291.

Schwarz, G. (1978). Estimating the dimension of a model. Annals of Statistics, 6(2), pags. 15-18.
}
\keyword{ package }
