% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg-package.R
\docType{package}
\name{StepReg-package}
\alias{StepReg}
\alias{StepReg-package}
\title{StepReg: Stepwise Regression Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Stepwise regression is a statistical technique used for model selection. This package streamlines stepwise regression analysis by supporting multiple regression types(linear, cox, logistic, poisson, gamma, and negative binomial), incorporating popular selection strategies(forward, backward, bidirectional, and subset), and offering essential metrics. It enables users to apply multiple selection strategies and metrics in a single function call, visualize variable selection processes, and export results in various formats. StepReg offers a data-splitting option to adress potential issues with invalid statistical inference and a randomized forward selection option to avoid overfitting. We validated StepReg's accuracy using public datasets within the SAS software environment. Additionally, StepReg features an interactive Shiny application to enhance usability and accessibility.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JunhuiLi1017/StepReg/issues}
}

}
\author{
\strong{Maintainer}: Junhui Li \email{junhui.li11@umassmed.edu} (\href{https://orcid.org/0000-0003-3973-1700}{ORCID})

Authors:
\itemize{
  \item Junhui Li \email{junhui.li11@umassmed.edu}
  \item Kai Hu \email{kai.hu@umassmed.edu}
  \item Xiaohuan Lu
  \item Wenxin Liu
  \item Lihua Julie Zhu
}

Other contributors:
\itemize{
  \item Sushmita N Nayak [contributor]
  \item Cesar Bautista Sotelo [contributor]
  \item Michael A Lodato [contributor]
}

}
\keyword{internal}
