% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\docType{package}
\name{StepReg-package}
\alias{StepReg-package}
\alias{StepReg}
\title{StepReg: An R package for stepwise regression analysis}
\description{
The package provides functions to perform Stepwise Regression using various algorithms, e.g., Forward Selection, Backward Elimination, and Bidirectional Elimination. It also supports the Best Subset method. Various types of models are supported such as "linear", "logit", "cox", and "poisson" regression. Users can choose from a wide range of selection criteria (stopping criteria) including AIC, AICc, BIC, HQ, and more.
}
\details{
\if{html}{\figure{logo.png}{options: style = "float: right"; alt = "logo"; width = "120"}}
\if{latex}{\out{\begin{center}}\figure{logo.png}\out{\end{center}}}

\tabular{ll}{ Package: \tab StepReg\cr Type: \tab Package\cr GitHub: \tab \url{https://github.com/JunhuiLi1017/StepReg}\cr}
}
\references{
1. \url{https://CRAN.R-project.org/package=StepReg}
}
\author{
\itemize{
\item Junhui Li \email{junhui.li11@umassmed.edu} (\href{https://orcid.org/0000-0003-3973-1700}{ORCID})
\item Kai Hu \email{kaihu@umassmed.edu}
\item Xiaohuan Lu
\item Kun Cheng
\item Sushmita Nayak
\item Cesar Bautista Sotelo
\item Michael Lodato
\item Robert H Brown
\item Lihua Julie Zhu
\item Wenxin Liu
}
\strong{Maintainer}
\itemize{
\item Junhui Li \email{junhui.li11@umassmed.edu} (\href{https://orcid.org/0000-0003-3973-1700}{ORCID})\cr
\item Kai Hu \email{kai.hu@umassmed.edu}
}
}
\keyword{package}
