% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg-package.R
\docType{package}
\name{StepReg-package}
\alias{StepReg}
\alias{StepReg-package}
\title{StepReg: Stepwise Regression Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package provides functions to perform Stepwise Regression using various algorithms, e.g., Forward Selection, Backward Elimination, and Bidirectional Elimination. It also supports the Best Subset method. Various types of models are supported such as "linear", "logit", "cox", and "poisson" regression. Users can choose from a wide range of selection criteria (stopping criteria) including AIC, AICc, BIC, HQ, and more.
}
\author{
\strong{Maintainer}: Junhui Li \email{junhui.li11@umassmed.edu} (\href{https://orcid.org/0000-0003-3973-1700}{ORCID})

Authors:
\itemize{
  \item Junhui Li \email{junhui.li11@umassmed.edu} (\href{https://orcid.org/0000-0003-3973-1700}{ORCID})
  \item Kai Hu \email{kai.hu@umassmed.edu}
  \item Xiaohua Lu
  \item Lihua Julie Zhu
  \item Wenxin Liu
}

Other contributors:
\itemize{
  \item Kun Cheng [contributor]
  \item Sushmita Nayak [contributor]
  \item Cesar Bautista Sotelo [contributor]
  \item Michael Lodato [contributor]
  \item Robert H Brown [contributor]
}

}
\keyword{internal}
