% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepReg.R
\docType{package}
\name{StepReg-package}
\alias{StepReg-package}
\alias{StepReg}
\title{An R package for stepwise regression analysis}
\description{
The package provides functions to perform Stepwise Regression using various algorithms, e.g., Forward Selection, Backward Elimination, and Bidirectional Elimination. Meanwhile, it also supports the Best Subset method. Multiple selection criteria (stopping criteria) are supported, including AIC, AICc, BIC, HQ, and more.
}
\details{
\tabular{ll}{ Package: \tab StepReg\cr Type: \tab Package\cr Version:
\tab 1.5.0\cr GitHub: \tab https://github.com/JunhuiLi1017/StepReg\cr Date: \tab 2023-01-23\cr }
}
\references{
1. [to be added]
}
\author{
Junhui Li, Kai Hu, Xiaohuan Lu, Kun Cheng, Sushmita Nayak, Cesar Bautista Sotelo, Michael Lodato, Robert H Brown, Lihua Julie Zhu, Wenxin Liu

Maintainers:\cr
Junhui Li <junhui.li11@umassmed.edu>\cr
Kai Hu <kai.hu@umassmed.edu>
}
\keyword{package}
