% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateUtils.R
\name{validateUtils}
\alias{validateUtils}
\title{Input validation}
\usage{
validateUtils(
  formula,
  data,
  type = c("linear", "logit", "poisson", "cox"),
  include = NULL,
  strategy = c("forward", "backward", "bidirectional", "subset"),
  metric = c("AIC", "AICc", "BIC", "CP", "HQ", "HQc", "Rsq", "adjRsq", "SL", "SBC",
    "IC(3/2)", "IC(1)"),
  sle = 0.15,
  sls = 0.15,
  sigma_value,
  test_method_linear = c("Pillai", "Wilks", "Hotelling-Lawley", "Roy"),
  test_method_logit = c("Rao", "LRT"),
  test_method_poisson = c("Rao", "LRT"),
  test_method_cox = c("efron", "breslow", "exact"),
  tolerance = 1e-06,
  weight = NULL,
  best_n = Inf,
  excel_name = NULL
)
}
\value{
Stop and exit if any error detected in input parameters
}
\description{
Validate the input parameters passed to stepwise()
}
\author{
Junhui Li, Kai Hu
}
