% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote.R
\name{vote}
\alias{vote}
\title{Vote for all models}
\usage{
vote(x, ...)
}
\arguments{
\item{x}{each dataframe from outputlist}

\item{...}{further parameters}
}
\value{
formatted dataframe
}
\description{
Votes for all models across all combinations of strategies and metrics
}
